REST Calculator
    - Fix package and proyect name.
    - Add new binary operations (Multiplication and Division).
    - Update readme for dummies (like me).
