+++ src/main/java/com/fabian/calculator/CalculatorApp.java
@@ -1,0 +1,11 @@
+package com.fabian.calculator;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication()
+public class CalculatorApp {
+    public static void main(final String[] args) {
+        SpringApplication.run(CalculatorApp.class, args);
+    }
+}
