+++ src/main/java/com/fabian/calculator/model/SubtractExpression.java
@@ -1,0 +1,8 @@
+package com.fabian.calculator.model;
+
+import java.math.BigDecimal;
+
+public class SubtractExpression extends AbstractBigDecimalExpression {
+    @Override
+    protected BigDecimal execute(final BigDecimal a, final BigDecimal b) { return a.subtract(b,getPrecision());}
+}
