+++ src/main/java/com/fabian/calculator/payload/input/Input.java
@@ -1,0 +1,12 @@
+package com.fabian.calculator.payload.input;
+
+import com.fabian.calculator.payload.Payload;
+
+public final class Input extends Payload {
+    @Override
+    public String toString() {
+        return "Input{" +
+                super.toString() +
+                '}';
+    }
+}
