+++ src/main/java/com/fabian/calculator/service/CalculatorService.java
@@ -1,0 +1,5 @@
+package com.fabian.calculator.service;
+
+public interface CalculatorService {
+    String process(final String expression);
+}
