+++ src/main/java/com/fabian/calculator/util/Constants.java
@@ -1,0 +1,16 @@
+package com.fabian.calculator.util;
+
+public final class Constants {
+
+    public static final String SUM_SIGN = "+";
+    public static final String MINUS_SIGN = "-";
+    public static final String MULTI_SIGN = "*";
+    public static final String DIV_SIGN = "/";
+    public static final String BLANK_STR = "\\u0020";
+
+    public static final String ERROR = "ERROR";
+    public static final String NOT_NUMERIC_OPERAND = "NON-NUMERIC OPERAND";
+    public static final String EXPRESSION_IS_NOT_COMPLETE = "EXPRESSION IS NOT COMPLETE";
+
+    private Constants(){}
+}
