--- src/main/java/com/sanitas/calculator/CalculatorApp.java
@@ -1,11 +1,0 @@
-package com.sanitas.calculator;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication()
-public class CalculatorApp {
-    public static void main(final String[] args) {
-        SpringApplication.run(CalculatorApp.class, args);
-    }
-}
