--- src/main/java/com/sanitas/calculator/model/AbstractBigDecimalExpression.java
@@ -1,30 +1,0 @@
-package com.sanitas.calculator.model;
-
-import com.sanitas.calculator.model.core.AbstractBinaryOperation;
-
-import java.math.BigDecimal;
-import java.math.MathContext;
-
-import static java.math.MathContext.DECIMAL128;
-
-public abstract class AbstractBigDecimalExpression extends AbstractBinaryOperation<BigDecimal> {
-
-    private MathContext precision;
-
-    protected AbstractBigDecimalExpression(final MathContext precision) {
-        this.precision = precision;
-    }
-
-    /**
-     * Default precision
-     */
-    protected AbstractBigDecimalExpression() {
-        this(DECIMAL128);
-    }
-
-    MathContext getPrecision() {
-        return precision;
-    }
-
-
-}
