--- src/main/java/com/sanitas/calculator/model/AddExpression.java
@@ -1,9 +1,0 @@
-package com.sanitas.calculator.model;
-
-import java.math.BigDecimal;
-
-public class AddExpression extends AbstractBigDecimalExpression {
-
-    @Override
-    protected BigDecimal execute(final BigDecimal a, final BigDecimal b) { return a.add(b);}
-}
