--- src/main/java/com/sanitas/calculator/model/core/Expression.java
@@ -1,11 +1,0 @@
-package com.sanitas.calculator.model.core;
-
-import java.util.Stack;
-
-/**
- * @autor FabianSR
- * @param <T>
- */
-public interface Expression<T extends Number> {
-    void interpret(final Stack<T> context);
-}
