--- src/main/java/com/sanitas/calculator/model/core/OperandExpression.java
@@ -1,13 +1,0 @@
-package com.sanitas.calculator.model.core;
-
-import java.util.Stack;
-
-public class OperandExpression<T extends Number> implements Expression<T> {
-    private T value;
-    public OperandExpression(final T value) {this.value = value;
-    }
-    @Override
-    public void interpret(Stack<T> context) {
-        context.push(value);
-    }
-}
