--- src/main/java/com/sanitas/calculator/payload/input/Input.java
@@ -1,12 +1,0 @@
-package com.sanitas.calculator.payload.input;
-
-import com.sanitas.calculator.payload.Payload;
-
-public final class Input extends Payload {
-    @Override
-    public String toString() {
-        return "Input{" +
-                super.toString() +
-                '}';
-    }
-}
