--- src/main/java/com/sanitas/calculator/service/CalculatorService.java
@@ -1,5 +1,0 @@
-package com.sanitas.calculator.service;
-
-public interface CalculatorService {
-    String process(final String expression);
-}
